backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg[binary]==3.2.7
python-dotenv==1.1.0
PyJWT==2.10.1
pydantic==2.11.3
